Add support for response files with command line arguments.
Response files that contain additional command line arguments may be
specified using the @<file> syntax similar to other Microsoft build
tools.

This enables over 8192 character command lines on Windows, which may be
necessary when testing a large number of assemblies.

Note that much of this code was lifted from the Roslyn repository, which
is copyright Microsoft but released on an Apache license.